+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - More options to let user customize the formatting. User can now opt to have custom pattern formatting. The user can also keep the options empty to skip formatting the patterns. The user can now also let the formatter auto-add headers when updating saved ones.
 - Grammar to detect address related operations and register usage.
 - An advanced parser to parse and gather more information about a pss source file.
+- There are new functions to use the advanced parser. These include a new goto definition provider, and an autocompletion provider.
+- There is a new hover provider to provide basic information about a hovered object.
 
 ### Fixed